+++ packages/components/psammead-social-embed/src/Canonical/providers.json
@@ -1,0 +1,9 @@
+{
+  "instagram": {
+    "script": "https://www.instagram.com/embed.js"
+  },
+  "twitter": {
+    "script": "https://platform.twitter.com/widgets.js"
+  },
+  "youtube": {}
+}
